Removing the dependency utilization feature
DWYU should have a well defined scope following the Unix philosophy:
"Do one thing and do it well"

The focus of DWYU is making sure the include statements in a targets
source code and the targets dependency list are matching.
Judging if a target should be restructured, because only a few of its
headers are used at a call site, does not directly match the DWYU focus.
Especially, given the usage of virtual includes makes the feature unreliable
either way.

If we want this feature back at some point, we should introduce a
separate tool which reuses DWYUs source code by making parts of DWYU
available as library.