Remove outdated README parts
The "word of warning" is not helpful. It can decrease a users trust into
the project. While it is true that some parts of DWYI might change, this
does not require a dedicated warning. Change is a natural process for a
software project. Breaking changes can be communicated with the version
number and release notes. What will most likely change in the future is
transparent due to the GitHub Issues and projects.

The roadmap has not much meaning given the GitHub Issues document what
is going on to happen with DWYU.