Disable temporarily coverage and pylint checks
Coverage is running on tests in the external dir, which we don't have
under control. Pylint has problems resolving imports between our files.
It seems we have to restructure how those tools are executed and configured.
We will do this in a separate step and accept for now the risk of quality
regressions.