Enable apply_fixes to work on custom bazel-bin directories
If a user uses a non fastbuild compilation mode or a custom output base
to generate the DWYU reports, the automatic deduction of the directory
containing the DWYU reports fails. Thus, provide the users with an option
to specify the directory where the apply_fixes tool should search for the
DWYU report files.