Enable the apply_fixes tool to utilize the convenience symlinks
The convenience symlink at the workspace root which points at the
bazel-bin directory is not a bullet proof approach to find the bazel-bin
directory. Users can remove it or rename it. Still, the symlink is created
by Bazel with the default setup and is a convenient way to reach the
bazel-bin directory matching the compilation mode of the previous commands.