Drop dataclass decorator from AvailableDependencies class
Python 3.6 does not provide the `dataclasses` module as part of the standard library.
We rename the attribute for the own headers, since `self` is a reserved
name which we should not use for a custom member.
Also we change the data type of the former self attribute, since using
the `AvailableDependency` class did not provide any value, but instead
obfuscated this is only a but a list of headers.