Apply fixes is aware of potential common errors
Even if buildozer does not return 0, this might not be a failure.
There are multiple return codes for various cases. Each has its own list
of potential causes. Not all of them are a real error, but maybe just
related to the limits on what buildozer can do due to the limited information
available to it.