Establish new apply_fixes acceptance testing framework
While working on new apply_fixes features it became clear that the old
approach is not flexible enough and is bad for test readability.

One class aiming at providing tools for all possible edge cases in
execution becomes cluttered. Also, while it seems elegant at first to be
able to specify what the test cases should do on a high level, it makes it
actually harder to understand what exactly is going on at a later time.

Thus, we now use dedicated files for each test, which implement the test
logic with support of generic helper functions. This eases understanding what
happens during a test and makes it easy to deviate from the established
pattern in single test cases.