Distinguish between invalid public and private includes
We want to offer an automatic fix for invalid includes, where we search for
a dependency which provides the header used in the target sources.
To be able to add the dependency correctly to either `deps` or
`implementation_deps` we have to know of the invalid include was used in a
public or private source file.