Simplify dependency discovery logic
We search only for dependencies potentially providing a header for an
include statement in the dependency tree of the target under inspection.
We work with the assumption that the code we analyze is compiling. Thus,
the target providing the required headers has to be in the transitive
dependencies.

We query all dependencies of the target under inspection and their headers
once and then analyze this ourselves without further query calls. This
eases testing and improves the performance as subprocess calls are quite
expensive.