Limit incompatibility flags with min and max Bazel version
We use several incompatibility flags to make sure the implementation does
not depend on deprecated behavior.
However, doing so via the .bazelrc file has the downside that it is limited
to flags which are available in all versions under test.
Thus, we prefer setting the incompatibility flags in the test script where
we can control them more fine grained. Also we should have all flags in a
single place. On top we add the missing feature of limiting the max
compatible Bazel version (e.g. when a flag is deprecated in a newer version).