Increase minimal Bazel version to 5.0.0
Previously we left this to prevent breaking for 4.x users which
theoretically still can use DWYU with Bazel 4.x. No breaking change was
introduced yet. Then again, we only test >= 5.0.0 and keeping this check
lower seems like an error on a first glance. Instead of adding a comment
explaining this we increase the minimal version. Adding a comment and
hoping we realize we introduce a breaking change with 4.x without testing
4.x anymore is just asking for creating a bug.