Add mypy dependency
We use type annotations in our Python code. However, without a tool
checking those, we rely on seeing all errors in the IDE analysis, if the
IDE supports typing at all.
Thus, we introduce mypy as test step to ensure our code adheres the typing
annotations

While doing so we restructure the setup macros. We cannot combine normal
user facing setup and dev setup into a single .bzl file. This worked until
now by accident, since skylib is both a public and a dev dependency. But
for dependencies which are purely for development, this would fail as the
load statements for development only dependencies cannot be resolved from a
user loading and executing the user facing setup macros.