Remove caching in CI job
The download of Bazel and the third_party artifacts is quite fast. We don't
profit much from caching it.
At the same time our bazel-out is quite large due to testing multiple
Bazel and Python versions and creating multiple output directories for
this to speed up local incremental testing.
Uploading this takes > 2 minutes. Given we rarely require multiple CI runs
on the same PR this is on average not worth it for speeding up potentially
CI runs on PR updates.