Support TreeArtifacts
rules using `ctx.actions.declare_directory(..)` to create a directory
containing an amount of generated code unknown at analysis phase is a
common pattern for code generators. Imagine for example a generator where
a provided configuration file determines which code is generated exactly.

Such a predeclared directory with unknown content is called `TreeArtifact`.
The content of a `TreeArtifact` cannot be analyzed in Starlark, due to the
analysis only being possible in the execution phase. Thus, we move the DWYU
logic partly into a Python script which is executed in the execution phase
to analyze various things we did before directly in the aspect.