Use lists for optional input to the aspect
Having a private and empty aspect config as impl detail for signaling that
no external config was provided is kinda hacky.
Using a list is the more elegant workaround for the limited aspect attr API
due to "empty" vs "not empty" matching closer to "nothing has been provided"
and "User provided a value".