Update rules_python to 0.25.0
This is a breaking change, as it raises the minimum Bazel version to 5.4.0.
Given Bazel 7.0.0 is released soon, expecting most users no longer relying
on 5.0.0 seems reasonable. Furthermore, we depend heavily on rules_python
which requires Bazel >= 5.4.0 for all its recent versions. Given this is
our main dependency and we aim at supporting bzlmod eventually, we require
access to recent rules_python versions.