Start using bzlmod
We can't use bzlmod with Bazel versions < 6.2.0.
Publishing DWYU to the central registry will be done in later steps. For now
we only enable using bzlmod while developing.
One downside of bzlmod is that rules_python no longer allows defining the
Python toolchain with patch level precision. As a consequence we have to
change the Python toolchains to the defaults defined by rules_python.