Soft vendor pcpp
We do not check in pcpp, but we fetch it via a standard http_archive call
instead of relying on rules_python and pip_parse.

pcpp is a pure Python module without transitive dependencies or platform
specific optimizations. Thus, using it without rules_python and pip_parse
is trivial.
Then again, Using rules_python and pip_parse to manage a dependency the user
requires as well is troublesome. It requires many dedicated steps in the
WORKSPACE setup. In the bzlmod case the code is more elegant, but it is
unclear how to best handle the toolchain dependency of pip_parse.
Therefore, we use a direct and simple approach to make pcpp available to
DWYU without burdening DWYU users with the complexity of pip_parse.