Replace aspect integration tests framework
The new test framework has the same design as used by the automatic fixes
test framework: Individual test are dedicated files.
This moves the test definitions closer to the code they are testing and also
allows custom logic and explanatory comments in the tests.

Also we decouple the test cases from the workspace bazelrc file to gain
more flexibility in using Bazel flags which are only available for some
Bazel versions.

Furthermore, we improve efficiency by:
- Combining some test cases which had no reason for separation besides the
  old test framework not supporting target patterns for multiple targets.
- using `--nolegacy_external_runfiles`