Add examples
It is a common best practice to have well documented examples which
showcase the usage of a project. Until now we pointed to some of the
integration tests for this. They are however a bad fit for this. Their
documentation is only the minimum to understand the test cases and often
they test edge cases not relevant for the user.