Migrate main workspace to bzlmod
All tests requiring an old Bazel version are located inside the integration
tests and examples which have their won sub workspaces.
Thus we can drop support for the legacy WORKSPACE setup in the main repo.
At the same time we simplify the main toolchain. It requires only a single
version.