apply_fixes can parse report files from a DWYU execution log
For large workspaces discovering the DWYU report files by crawling the bazel-out
directory can be quite slow due to an enormous amount of files and directories
being present.
To work around this, we enable the apply_fixes script to parse a log file
containing the command line output of executing the DWYU aspect. This execution
log is then parsed and the DWYU report paths deduced.