Remove mypy aspect
The project providing the aspect does not offer a proper modern bzlmod
integration. Contribution is difficult due to the maintainer having
almost no time.
We do not really need a bazelized solution for this. We already
use the pre-commit framework and it offers as well a mypy integeation.
On top, we can more easily stop ignoring import errors with the
pre-commit based mypy integration.