Use .bazelignore to allow testing '//...'
We did not do so in the past due to how mypy discovered Python files with
Python in the past. Now that we use mypy without Bazel via pre-commit
this restriction no longer applies.
We prefer maintaining a .bazelignore file instead of a detailed test command.
This prevents us missing tests.