Experimental support for setting the '__cplusplus' macro
`__cplusplus` is a macro defined by the compiler specifying if C++ is used to compile the file and which C++ standard is used.
`__cplusplus` typically is not passed on the compiler command line as defined value, but set internally by the preprocessor while processing a file.
Thus, DWYU does not know about it when performing the processing to resolve preprocessor statements.

To work around this, DWYU we can set `__cplusplus` based on a heuristic.
This heuristic is for now a best guess and thus the feature is experimental without any guarantee for stability.