refctor: Stop using poetry
Using poetry merely for running pre-commit is overkill. Especially, given
the pre-commit GitHub action either way ignored our setup and
downloaded and used the latest pre-commit version.
pre-commit is an established tool with a stable behavior. All users can
be expected to install it for themselves. Furthermore, we can ensure a
fairly recent version is used via the pre-commit config.