build(deps): Update rules_cc to 0.1.0 for WORKSPACE
rules_cc has an annoying dependency on protobuf without providing a WORKSPACE
macro allowing the user to automatically fetch the matching version. This protobuf
dependency seems to be undesired, as rules_cc version 0.1.0 drops it.
We raise the version only for WORKSPACE and not for bzlmod, as with bzlmod the
protobuf dependency is automatically handled in the background. rules_cc version
0.15.0 is sufficient for us regarding C++ features and with bzlmod one should
define the minimum required version, not the latest greatest compatible one.