test(examples): Drop hermetic Python toolchain
Setting up the hermetic toolchains is superfluous noise in the examples. Also,
it requires maintaining which rules_python version we use there. Plus, loading
rules_python in the examples on top of the DWYU dependencies might hide problems
with the rules_python version DWYU depends on.
Given we never had any issue with supporting multiple Python versions, we believe
using the host interpreter in the examples is fine.