test(aspect): Test setup uses rules_python version requested by DWYU
We abandon testing newer rules_python versions as the latest versions are not
compatible to Bazel 6 anymore. Also testing newer rules_python via the aspect
integration tests was always kinda hacky.
If we want to test the latest and greatest versions of our dependencies we
should setup dedicated tests for that similar to the dedicated
workspace_integration tests.