build: Switch to Bazel 8.0.1 for development and default testing
We drop the dev dependency to stardoc, which we do not actually need.

We stop using `--nolegacy_external_runfiles` as it is the default with Bazel 8.
Strictly speaking this reduces the performance in some aspect integration tests.
But given the difference is minimal, we prefer an overall clean and consistent
setup not mentioning this flag.