feat(aspect): Deduce all headers provided by the CC toolchain
Until now we hard coded a list of standard library and system headers.
While this works in most cases, it is far from perfect. One easily overlooks
system headers by this approach.

The new behavior is for now an opt-in, as we might have overlooked issues
while testing. Bazel CC toolchains can be written with 2 different APIs and
either only reference host artifacts, a mix of host and hermetic artifacts
or a fully hermetic toolchain. Given this complexity, we want to give users
time to file bug reports in case of systematic design flaws.
While we start slow and careful, we definitely aim at making this the default
behavior eventually. Since we are positive this will become the new
default implementation, we do not mark this as experimental.