doc(aspect): Document limitation when using copts for include paths
We do not plan to support this use case anytime soon. The usage of copt
and similar options implies compiler specific syntax. While, most people
use gcc/clang, anybody is free to use any other compiler. Introducing the
complexity of compiler specific behavior for something which users can also
do via a portable Bazel API is for now not considered to be worth it.