feat(aspect): First bite of switching implementation to C++
We received only few functional defects in the last months, but several
issues regarding performance were raised. We could mitigate some of those
due to community contributions fixing some bad complexity in our algorithms.
Still, performance is lacking in large projects or in certain edge cases
for complex code.

We experimented with addressing this by using the CC toolchain preprocessor.
Unfortunately, this failed as its impossible to obtain the information
about direct includes this way. We might try again using a
C++ preprocessor library to do the same as we do today in Python, but
with a more efficient implementation. This is however work for a future
patch, as it is a quite big change.
One of the reasons to use Python was to ease coming up with a prototype for
this projects idea and to facilitate fast iterations. By now the core
structure is mostly stable and we have a good test coverage. Thus, now we
are ready to exchange the implementation with something new.

For now, we take a first bit of switching to a C++ based implementation
with the tool processing the CompilationContext values of the CC targets
under inspection to unroll the values from TreeArtifacts (e.g. header files).
This is a small piece of code which is easy to migrate, but at the same time
offers considerable performance improvements. This tool is executed quite
often (once for each direct dependency of all CC targets under inspection).
Thus, the significant overhead of executing something with the hermetic
Python toolchain (thousands of symlinks for each action) is measurable.
By using a C++ based tool we get rid of the Python toolchain overhead.
A synthetic benchmark running DWYU on a project with many trivial CC
targets shows a speedup greater factor 2. If the code of the CC
targets grows in complexity, this speedup shrinks as the code analysis
then requires the majority of the time. So, in real projects no dramatic
speedup can be expected. Still, it improves things and is a first step.
Even when using the non hermetic host Python interpreter, the new C++ based
implementation is significantly faster than the old Python based one. Thus,
showing that not only the hermetic Python toolchain overhead is an issue,
but Python itself is slower compared to C++ for the specific tasks we perform,
which does not come as a surprise.

DWYU does not provide its own CC toolchain. The users CC toolchain is used
to compile the C++ tool. The minimum required C++ version is C++11.

For now users have to opt into the C++ implementation in case we overlooked
some problem or side effect. But the intention is to make the C++ impl
the default DWYU behavior soonish.
Activate the new implementation by using the new factory option
`use_cpp_implementation`.