test(aspect): More robust implementation for set_cpp_standard test
We are not compiling the code. Thus, it is an implementation detail if a
missing header for an include statement actually causes an error. We could
also ignore such cases as part of a heuristic.
Therefore, we choose a test design which is based on a core DWYU principle
for which we will always fail.