feat(aspect): Extend C++ impl by a boost wave based preprocessor
This is the next step in migrating the whole DWYU aspect implementation to
C++ for increased performance.
The preprocessor is the slowest part of of the legacy Python DWYU
implementation. Thus, this step provides a significant performance boost.
Analyzing code is now up to 6 times faster. The exact speedup depends on
your code complexity an size. Easy to process code profits less.
Also, the new approach analyzes each file individually. Thus, for targets
with many files and many jobs being available (e.g. remote execution)
You can see significant speedups as well.

This is not yet a feature complete preprocesing step. It does not support
the "no preprocessor" mode and we want to make it configurable with various
quality and speed trade offs for preprocessing. Given, users still are able
to use the Python based implementation, it is fine to merge intermediate
steps. Especially, while features are missing, the relevant tests are all
green.