feat(aspect): Deprecate feature 'experimental_set_cplusplus'
This is an edge case feature which can be resolved without any dedicated
logic, as shown in the `define_macros` example. The alternative is even
more powerful due to working for all macros, not just `__cplusplus`.

We do not intend to maintain the custom logic to determine the desired
C++ standard any further.

If it turns out there are use cases where utilizing `--cxxopts` is not
feasible, we will think about another approach where one can provide DWYU
with a list of custom defines via a support target through a provider.