fix(aspect): Properly process non trivial implementation_deps
Because of bug https://github.com/bazelbuild/bazel/issues/19663 we are
missing the information from `implementation_deps` to properly process them
if they are from an external workspace or modifying their include paths.
This was not discovered until now as `implementation_deps` to workspace
internal targets worked if they did not manipulate the include paths. In
such a case we could still resolve the headers through the include path
pointing to the workspace root, which is essentially always there.