feat(aspect): Add "no_preprocessing" to the C++ implementation
In our benchmarks the `no_preprocessor` mode is up to 8x faster for complex
code. However, for trivial code there is almost no difference and for medium
complexity code we see a speedup around 3x.
We still recommend using the preprocessor by default, as it yields the
more correct results and especially with the new C++ based implementation
is overall decently fast.