build: Increase minimum Bazel version to 7.2.1
We cannot support Bazel 9 and Bazel 6 at the same time.
Our new minimum Bazel version is 7.2.1 as some of our dependencies for
C++ implementation demand this as minimum version.

For now we drop the Python 3.9 test support. Python 3.9 is either way end
of life and we do not change the Python implementation anymore. As soon as
the C++ based implementation is the default, we will reduce the tested
Python versions even more and stop officially supporting Python < 3.10.