refactor: Ensure C++ code quality with compiler warnings
We are not using a dedicated config so we see quality problems immediately
during development and do not have to introduce an extra CI step for the
compiler warnings.
We can even remove the CI step building the code with various C++ standards
as we can esnure C++11 compliance with compiler warnings.