feat(aspect): Add a C++ based header file analysis
We move the last piece of the DWYU implementation from Python to C++ to
offer a fully C++ based implementation of the DWYU aspect.

There are no user facing changes.

However, internally we restructure our approach. The preprocessing step is
now providing the sandbox paths to the included files. This information is
for free, as the preprocessor either way has to locate the headers to open
them for preprocessing. On Bazel side it is also trivial to get the
sandbox paths of all the headers provided by the various dependencies.
This simplifies a lot the logic for analyzing the include statements in the
target under inspection. We no longer search files ourselves in the
sandbox, which was the most expensive thing we did in the analysis. The
analysis step is now focused on comparing the header files from the
dependencies to the included files reported by the preprocessing step.

Another big internal change is that the allow list of C/C++ standard
library headers and the code to automatically look up all headers provided
by the Bazel CC toolchain are unused by the new DWYU aspect implementation.
While preprocessing, we omit the CC Toolchain headers from the sandbox.
Only the headers of all dependencies a available. The preprocessing
silently skips all headers it cannot localize under the assumption that
those have to be headers from the CC toolchain. For one, this improves the
speed and stability of the preprocessing, as the CC toolchain headers are
quite hard to peprocess. Another benefit is, that the ignoring of the
standard library headers is for free without the analysis step having to
compare detected includes to a list of CC toolchain headers.